Add wanted_events to build an inotify filter mask
If you only want to handle a subset of event types, allow
passing an array ref of wanted_events which will be
used to build a bitmask of events to listen for when
creating the inotify watcher.

Conceptually this is similar to the filter predicate, but
because it's passed to the inotify call, it filters events
at the kernel level which can avoid queue overflows if
a large number of events you aren't interested in are
generated.

Additionally, lets protect against queue overflow events by
just adding an on_overflow handler that does nothing.
This is better than reporting a BUGBUG that dumps the
entire object structure.